requirements.txt
openai==0.27.5
python-dotenv
python-telegram-bot